allow items to be picked up during interaction and delete them off the ground

+++ random.hpp
@@ -3,6 +3,6 @@
 
 std::mt19937_64 mt(0);
 
-int getRandomNum() {
+int64_t getRandomNum() {
 	return mt();
 }
